Add driver interface
The driver interface handles wiimote I/O and connects it with the
uinput interface. The uinput interface can be used from userspace
so the specification is located in the new library directoy. The
library headers can be installed into the system and then used by
applications to access the input device.