Add input support to wiimote library
Add input parser which reads the /dev/input/eventX device of the wiimote
and caches the received values. This allows the user to read the current
state of each wiimote at any time without querying the device.

Also the declaration for device queries are added but not implemented
yet. They will allow to read the sysfs attributes of each wiimote and
retrieve state values which are not sent through the input event
interface.