Add remaining DRM input reports to PROTOCOL doc
Add remaining DRM input reports without documentation to PROTOCOL doc.
The documentation is fairly simply and may be added later.

The PROTOCOL file now contains documentation for all reports that are
supported by the wii as exposed by its report descriptor. There may be
more reports which are not included in the report descriptor, though.