Add dummy compatibility library
Since adopting a new API is always hard we now try to emulate the two
most commonly used Wii Remote APIs under linux. This is cwiid and
wiiuse. The current implementation is just a dummy but this will
eventually become a full ABI compatible implementation of both libraries
based on libxwiimote.so.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>