compat: remove compatibility interfaces
Cwiid and wiiuse have a horrible API design. It would take months to get the
compatibility library right and then we would end up with support for two more
APIs which no-one want to use. Instead, we should help porting old applications
to the new interface and everyone will be happy.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>