lib: fix monitor udev device filter
The xwii_monitor recently discovered devices by the "event" files. However, we
have multiple event files per HID device so we now directly discover only the
HID device. This prevents the monitor from returning a device multiple times.

As a side effect, the device might not be fully set up when the parent HID
device is broadcasted by udev. However, that's up to the user to handle as we
only notify him of new devices, not of the current state of the remote device.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>