xwiikeymap: sleep short time before setting up device
UDev reports when the parent HID device is set up. However, the child devices
like input-devices or led-devices may need some more time. Therefore, we sleep
for 10ms before opening the device.

We can ignore the delay even though we are unthreaded as it is really short.
However, in the future we should use more appropriate methods here.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>