lib: new xwii_iface_poll() function
In later versions we might need to send pending requests to the Wii Remote
device. To do this in one function together with reading events we introduce the
new poll function. It currently does the same as *_read() but has other
semantics which may change in the future.

To avoid big API changes later we introduce it now even though we don't need it,
yet.

Signed-off-by: David Herrmann <dh.herrmann@googlemail.com>