xwiishow: print visual accelerometer data
We now print 3D coordinate axes to visualize the accelerometer data. It
makes it a lot easier to see what the data represents.

We use pow(val, 1/4) to get smoother values around the origin. It is
quite impossible to reach the maximum accelerometer values for a longer
time so we concentrate more on the origin.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>