xwiishow: freeze screen with 'f' hotkey
We freeze the screen when the user presses 'f'. If it is pressed again, we
continue as normal. This is useful to get snapshots for debugging.

Note that this breaks key events as we do not refresh them after unfreeze,
so they may be marked as "pressed" if you released them during the freeze.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>