Add normalization and calibration for MotionPlus
We now allow the API users to normalize MP input. Automatic recalibration
is also implemented to suppress small data errors.

MP has the problem that it's data is in a range +/- 100 which is really
hard to come-by in applications using it. Therefore, we apply a varying
calibration value and provide an absolute offset as normalization.

(minor coding-style and API fixes)
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>