lib: add hotplug detection and event forwarding
We now provide xwii_iface_watch() which enables or disables hotplug
detection. We do this via udev_monitor and try to detect any kind of
hotplug event.

Whenever a hotplug event occurs, we send an XWII_EVENT_WATCH to the
caller.

Note that this detects generic hotplug events but does not provide any
additional information to the caller. As hotplug events are supposed to
be rare and only provide limited information, we require a caller to
re-enumerate all interfaces of a device to get more information. Sysfs
makes this pretty simple so there is no need for us to provide additional
hotplug payloads which describe what changed (which wouldn't be possible,
anyway).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>