lib: return XWII_EVENT_WATCH on EPIPE
Whenever an FD is closed, we now return XWII_EVENT_WATCH instead of
ENODEV so users can properly react to it in a generic way.

Note that this breaks the semantics of xwii_iface_poll(). However, the
next libxwiimote release will be ABI-incompatible, anyway, so we use
this to fix up this API as well.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>