lib: reread device-nodes on hotplug events
We need to close all devices that got removed during an hotplug event and
we need to re-read all new interfaces.

If we don't do this, a single extension unplug can cause up to XWII_IF_NUM
hotplug events sent to the caller, which is not what we want.

This also fixes a bug were we didn't correctly read new device-nodes that
were hotplugged but required the caller to recreate the state.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>