lib: remove old deprecated API
We break backwards-compatiblity with the next release so use this to also
drop old API entry-points and constants.

With the rewrite of the kernel driver we get a fully-hotplug capable
infrastructure so applications will have to adopt to it. But we still
support the old kernel driver the same way we did before! The API that
is now removed was never really used.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>