lib: open ifaces in a row regardless of errors
If multiple interfaces are opened in a single call, we shouldn't stop
on a single error. Instead, users are inclined to call xwii_iface_opened
to get a list of opened devices if xwii_iface_open() returned an error.

This avoids hard-coding the iface-order into the semantics of
xwii_iface_open() and handles them regardless of the type.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>