Bump version to 2
This is the xwiimote-2 release. We changed version numbers to integers and
skipped xwiimote-1 for compatibility reasons. We provide proper
symbol-versioning and libtool versions so compatibility with older
releases will be provided.

Note that this release breaks some APIs so applications will be required
to be recompiled.

See git-log for a list of changes. This basically adds support for the new
features introduced into hid-wiimote with linux-3.11 (to be released late
September).

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>