res: add example udev permission setup
We need to adjust LED attribute permissions to allow non-root users to
access wiimote LEDs. Attached is an example udev rule which does that.
We have to be careful not to recursively follow any links with the "chgrp"
and "chmod" calls, hence it's not installed by default. Feel free to copy
it into your system.

Reported-by: Aubrien
Signed-off-by: David Herrmann <dh.herrmann@gmail.com>