res: whitelist pro/classic controllers for X11
The gamepad extensions are independent of the core wiimote and can be
safely whitelisted as they follow the kernel's gamepad rules. There is no
reason to make these not work by default.

xf86-input-xwiimote doesn't make use of extensions, yet. And for classic
controllers there is no reason why it every would do so.

Signed-off-by: David Herrmann <dh.herrmann@gmail.com>